+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+#### 1.0.2 (2019-06-13)
+
+* Fix BC layer for symfony/config 4.1 and older.
+* Deprecated `oka_wsse_authentication.user_class` configuration values.
+
 #### 1.0.1 (2019-06-12)
 
 * Fixed symfony component requirements.